# web/openapi.json
{
  "components": {
    "schemas": {
...
      "ClickFeedback": {
        "properties": {
          "context": {
            "additionalProperties": true,
            "title": "Context",
            "type": "object"
          },
          "event_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Event Id",
            "type": "string"
          },
          "external_product_id": {
            "title": "External Product Id",
            "type": "string"
          },
          "impression_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impression Event Id"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "request_id",
          "external_product_id",
          "position"
        ],
        "title": "ClickFeedback",
        "type": "object"
      },
      "ConversionFeedback": {
        "properties": {
          "context": {
            "additionalProperties": true,
            "title": "Context",
            "type": "object"
          },
          "event_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Event Id",
            "type": "string"
          },
          "external_product_id": {
            "title": "External Product Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "event_id",
          "request_id",
          "external_product_id"
        ],
        "title": "ConversionFeedback",
        "type": "object"
      },
...
      "FeedbackResponse": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "duplicate": {
            "title": "Duplicate",
            "type": "boolean"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "feedback_type": {
            "title": "Feedback Type",
            "type": "string"
          },
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "feedback_type",
          "accepted",
          "duplicate",
          "received_at"
        ],
        "title": "FeedbackResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImpressionFeedback": {
        "properties": {
          "context": {
            "additionalProperties": true,
            "title": "Context",
            "type": "object"
          },
          "event_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Event Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecommendationItem"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "request_id",
          "items"
        ],
        "title": "ImpressionFeedback",
        "type": "object"
      },
...
      "RecommendationItem": {
        "properties": {
          "external_product_id": {
            "title": "External Product Id",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          }
        },
        "required": [
          "external_product_id",
          "position"
        ],
        "title": "RecommendationItem",
        "type": "object"
      },
...
      "RecommendationRequest": {
        "properties": {
          "context": {
            "$ref": "#/components/schemas/RecommendationContext"
          },
          "exclude_product_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "title": "Exclude Product Ids",
            "type": "array"
          },
          "fallback_allowed": {
            "default": true,
            "title": "Fallback Allowed",
            "type": "boolean"
          },
          "request_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "top_n": {
            "default": 10,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top N",
            "type": "integer"
          },
          "user_id": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "title": "RecommendationRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "GraphRec API",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/feedback/clicks": {
      "post": {
        "operationId": "submit_click_feedback_v1_feedback_clicks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClickFeedback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Click Feedback",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/v1/feedback/conversions": {
      "post": {
        "operationId": "submit_conversion_feedback_v1_feedback_conversions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversionFeedback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Conversion Feedback",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/v1/feedback/impressions": {
      "post": {
        "operationId": "submit_impression_feedback_v1_feedback_impressions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpressionFeedback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Impression Feedback",
        "tags": [
          "recommendations"
        ]
      }
    },
...
    "/v1/recommendations": {
      "post": {
        "description": "Serve Top-N recommendations and record what it took to serve them.",
        "operationId": "get_recommendations_v1_recommendations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recommendations",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/v1/recommendations/session": {
      "post": {
        "operationId": "get_session_recommendations_v1_recommendations_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session Recommendations",
        "tags": [
          "recommendations"
        ]
      }
    },
...
  }
}